Separate concerns: graph building, simplification, rendering and demo/notebook..
