+++ flatland/core/graph/grid_to_graph.py
@@ -24,20 +24,12 @@ TODO illustration simplification and edge cases (pun intended): non-facing switc
 TODO tests for conversion
 """
 from collections import defaultdict
-from typing import List, Tuple
 
-import matplotlib.pyplot as plt
 import networkx as nx
-import numpy as np
-from attr import attrs, attrib
 
 from flatland.core.grid.grid4_utils import get_new_position
 from flatland.core.transition_map import GridTransitionMap
-from flatland.envs.rail_env_utils import env_creator
 from flatland.envs.rail_env import RailEnv
-from flatland.utils.rendertools import RenderTool
-
-GridNode = Tuple[int, int, int]
 
 
 # TODO naming?
@@ -70,119 +62,3 @@ class GraphTransitionMap:
     @staticmethod
     def fromRailEnv(env: RailEnv) -> "GraphTransitionMap":
         return GraphTransitionMap(GraphTransitionMap.grid_to_digraph(env.rail))
-
-
-@attrs
-class DecisionPointGraphEdgeData:
-    path = attrib(type=List[GridNode])
-
-
-# TODO naming: decision point graph, simplified graph, collapsed graph, ....?
-class DecisionPointGraph:
-    def __init__(self, g: nx.MultiDiGraph):
-        self.g = g
-
-    @staticmethod
-    def _explore_branch(g: nx.DiGraph, u: GridNode, v: GridNode) -> List[GridNode]:
-        branch = [u, v]
-        successors = list(g.successors(v))
-        assert len(successors) > 0
-        while len(successors) == 1:
-            successor = successors[0]
-            branch.append(successor)
-            successors = list(g.successors(successor))
-            assert len(successors) > 0
-        return branch
-
-    @staticmethod
-    def fromGraphTransitionMap(gtm: GraphTransitionMap) -> "DecisionPointGraph":
-        g = nx.MultiDiGraph()
-
-        # find decision points (ie. nodes with more than one successor (=neighbor in the directed graph))
-        micro = gtm.g
-        decision_nodes = {s for s in micro.nodes if len(list(micro.successors(s))) > 1}
-
-        # add edge for
-        for dp in decision_nodes:
-            for n in micro.successors(dp):
-                branch = DecisionPointGraph._explore_branch(micro, dp, n)
-                g.add_edge(branch[0], branch[-1], d=DecisionPointGraphEdgeData(path=branch))
-        return DecisionPointGraph(g)
-
-
-# TODO separation of concerns: graph building and rendering...
-delta = 0.2
-offsets = {
-    # N
-    0: [0.5, -delta],
-    # E
-    1: [-delta, -0.5],
-    # S
-    2: [-0.5, delta],
-    # W
-    3: [delta, 0.5]
-}
-
-
-def _add_flatland_styling(env: RailEnv, ax):
-    env_renderer = RenderTool(env)
-    img = env_renderer.render_env(show=False, frames=True, show_observations=False, show_predictions=False, return_image=True)
-    ax.set_ylim(env.height - 0.5, -0.5)
-    ax.tick_params(left=True, bottom=True, labelleft=True, labelbottom=True)
-    ax.set_xticks(np.arange(0, env.width, 1))
-    ax.set_yticks(np.arange(0, env.height, 1))
-    # TODO image does not fill extent entirely - why?
-    ax.imshow(np.fliplr(np.rot90(np.rot90(img))), extent=[-0.5, env.width - 0.5, -0.5, env.height - 0.5])
-    ax.set_xticks(np.arange(-0.5, env.width + 0.5, 1), minor=True)
-    ax.set_yticks(np.arange(-0.5, env.height + 0.5, 1), minor=True)
-    ax.tick_params(left=True, bottom=True, labelleft=True, labelbottom=True)
-    ax.grid(which="minor")
-
-
-# TODO move to notebook
-if __name__ == '__main__':
-    env = env_creator()
-
-    micro = GraphTransitionMap.grid_to_digraph(env.rail)
-    gtm = GraphTransitionMap(micro)
-    decision_point_graph = DecisionPointGraph.fromGraphTransitionMap(gtm)
-    collapsed = decision_point_graph.g
-    fig, axs = plt.subplots(1, 2)
-
-    micro1 = nx.subgraph_view(micro, filter_edge=lambda u, v: len(list(micro.successors(v))) == 1)
-    nx.draw_networkx(micro1,
-                     pos={(r, c, d): (c + offsets[d][1], r + offsets[d][0]) for (r, c, d) in micro1},
-                     ax=axs[0],
-                     node_size=2,
-                     with_labels=False,
-                     arrows=False
-                     )
-    micro2 = nx.subgraph_view(micro, filter_node=lambda v: len(list(micro.successors(v))) == 2)
-    nx.draw_networkx(micro2,
-                     pos={(r, c, d): (c + offsets[d][1], r + offsets[d][0]) for (r, c, d) in micro2},
-                     ax=axs[0],
-                     node_size=8,
-                     node_color="red",
-                     with_labels=False,
-                     )
-    micro3 = nx.subgraph_view(micro, filter_edge=lambda u, v: len(list(micro.successors(v))) == 2)
-    nx.draw_networkx(micro3,
-                     pos={(r, c, d): (c + offsets[d][1], r + offsets[d][0]) for (r, c, d) in micro3},
-                     ax=axs[0],
-                     arrows=True,
-                     node_size=1,
-                     with_labels=False
-                     )
-
-    nx.draw_networkx(collapsed,
-                     pos={(r, c, d): (c + offsets[d][1], r + offsets[d][0]) for (r, c, d) in collapsed},
-                     ax=axs[1],
-                     node_size=2,
-                     with_labels=False
-                     )
-    _add_flatland_styling(env, axs[1])
-    _add_flatland_styling(env, axs[0])
-
-    axs[0].set_title('micro')
-    axs[1].set_title('collapsed')
-    plt.show()
